fixed pagination issue [PT #179188463]
when the displayed list length matches one of the options in the pagination size selector, it caused
"Warning: Encountered two children with the same key, `5`. Keys should be unique so that components maintain their identity across updates. Non-unique keys may cause children to be duplicated and/or omitted — the behavior is unsupported and could change in a future version."
Apparently MUI forgot to put a unique key on it...