config setting for webpack5 [PT #184553978]
ERROR in ./node_modules/fetch-mock/esm/client.js 661:18-25
Module not found: Error: Can't resolve 'process/browser' in '/Users/uqldegro/git/homepage-react/node_modules/fetch-mock/esm'
Did you mean 'browser.js'?
BREAKING CHANGE: The request 'process/browser' failed to resolve only because it was resolved as fully specified
(probably because the origin is strict EcmaScript Module, e. g. a module with javascript mimetype, a '*.mjs' file, or a '*.js' file where the package.json contains '"type": "module"').
The extension in the request is mandatory for it to be fully specified.
Add the extension to the request.